reorganised and renamed project structure, debug button bug
