+++ app/src/main/java/com/sdp13epfl2021/projmag/activities/ProjectFormActivity.kt
@@ -1,0 +1,93 @@
+package com.sdp13epfl2021.projmag
+
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
+import android.os.Bundle
+import android.provider.MediaStore
+import android.widget.Button
+import android.widget.MediaController
+import android.widget.VideoView
+import androidx.appcompat.app.AppCompatActivity
+import com.sdp13epfl2021.projmag.activities.tagsSelectorActivity
+
+
+class Form : AppCompatActivity() {
+    private val REQUEST_VIDEO_ACCESS = 1
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_form)
+        val addVideoButton: Button = findViewById(R.id.add_video)
+        val tagButton : Button = findViewById(R.id.addTagsButton)
+
+        addVideoButton.setOnClickListener {
+            val intent = Intent(Intent.ACTION_PICK, MediaStore.Video.Media.EXTERNAL_CONTENT_URI)
+            startActivityForResult(
+                Intent.createChooser(intent, "Select Video"),
+                REQUEST_VIDEO_ACCESS
+            )
+        }
+
+        tagButton.setOnClickListener {
+            switchToTagsSelectionActivity()
+        }
+    }
+
+
+    /**
+     * This function is called after the user comes back
+     * from selecting a video from the file explorer
+     */
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (resultCode == Activity.RESULT_OK && requestCode == REQUEST_VIDEO_ACCESS) {
+            if (data?.data != null) {
+                // THIS IS THE VID URI
+                val selectedVidURI = data.data
+
+
+                val playVidButton = findViewById<Button>(R.id.play_video)
+                val vidView = findViewById<VideoView>(R.id.videoView)
+                val mediaController = MediaController(this)
+
+                FormHelper.playVideoFromLocalPath(
+                    playVidButton,
+                    vidView,
+                    mediaController,
+                    selectedVidURI!!
+                )
+            }
+        }
+    }
+
+
+    fun switchToTagsSelectionActivity(){
+        //why do i need to do the :: class.java to make it work
+        System.out.println("helllloooooooooooooo")
+        val intent = Intent(this, tagsSelectorActivity::class.java)
+        startActivity(intent)
+    }
+}
+
+class FormHelper() {
+    companion object {
+        public fun playVideoFromLocalPath(
+            playVidButton: Button,
+            vidView: VideoView,
+            mediaController: MediaController,
+            uri: Uri
+        ) {
+            playVidButton.isEnabled = true
+            playVidButton.setOnClickListener {
+                vidView.setMediaController(mediaController)
+                vidView.setVideoURI(uri)
+                vidView.start()
+            }
+        }
+
+
+    }
+
+
+}
