+++ app/src/main/res/layout/activity_c_v_creation.xml
@@ -4,6 +4,6 @@
     android:id="@+id/cv_pager"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".curriculumvitae.CVCreationActivity">
+    tools:context=".activities.CVCreationActivity">
 
 </androidx.viewpager2.widget.ViewPager2>
